requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
asgiref==3.7.2
asteroid-filterbanks
attrs==23.2.0
audioread==3.0.1
Authlib==1.3.0
boto3==1.34.34
botocore==1.34.34
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
DateTime==5.4
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
Django==4.2.9
django-cors-headers==4.3.1
django-environ==0.11.2
docopt==0.6.2
docx==0.2.4
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2024.3.1
grpcio==1.60.1
grpcio-health-checking==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
httpcore==1.0.3
httpx==0.27.0
huggingface-hub==0.22.2
HyperPyYAML==1.2.2
idna==3.6
img2pdf==0.5.1
importlib-resources==6.1.1
ipynb-py-convert==0.4.6
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
julius
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
lightning
lightning-utilities==0.11.2
llvmlite==0.42.0
lxml==4.9.4
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==1.26.2
ocrmypdf==15.4.4
omegaconf==2.3.0
openai==1.12.0
openpyxl==3.1.2
optuna==3.6.1
packaging==23.2
pandas==2.1.4
pdfminer.six==20231228
pikepdf==8.11.2
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
pooch==1.8.1
primePy==1.3
protobuf==4.25.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
pyspellchecker==0.7.3
python-dateutil==2.8.2
python-docx==1.1.0
pytorch-lightning
pytorch-metric-learning
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
reportlab==4.1.0
requests==2.31.0
retrying==1.3.4
rich==13.7.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
semver==3.0.2
sentencepiece==0.2.0
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain
SQLAlchemy==2.0.29
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
textblob==0.17.1
threadpoolctl==3.2.0
tiktoken==0.6.0
torch-audiomentations
torch-pitch-shift
torchaudio
torchmetrics
torchvision
tqdm==4.66.1
typer==0.12.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
validators==0.22.0
virtualenv==20.25.0
weaviate-client==4.5.5
wheel==0.42.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.2
